pyproject.toml
[project]
name = "personal-memory-assistant"
version = "1.1.0"
description = "A local-first Personal Memory Assistant with hybrid RAG search"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
    "pywebview>=5.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "tenacity>=8.2.0",
    "sse-starlette>=1.6.0",
]

[project.scripts]
pma = "__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "S101",  # allow assert in tests
    "S603",  # subprocess calls are fine here
    "T201",  # allow print in CLI scripts
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
